main: split conifg_http_get/post.c
